refactor: 입력받은 닉네임을 이용해 Crew 객체를 찾는 역할을 CrewGroup이 수행하도록 리팩터링
