+++ src/main/java/attendance/model/CrewGroup.java
@@ -10,16 +10,18 @@ public class CrewGroup {
         this.crews = Set.copyOf(crews);
     }
 
-    public boolean contains(String nickname) {
-        return crews.stream()
-                .anyMatch(crew -> crew.isEqualsNickname(nickname));
-    }
-
     public boolean contains(Crew crew) {
         return crews.contains(crew);
     }
 
     public Set<Crew> getCrews() {
         return Set.copyOf(crews);
+    }
+
+    public Crew findCrewByNickname(String nickname) {
+        return crews.stream()
+                .filter(crew -> crew.isEqualsNickname(nickname))
+                .findFirst()
+                .orElseThrow(() -> new IllegalArgumentException("등록되지 않은 닉네임입니다."));
     }
 }
